python_ai/requirements.txt
flask
flask-cors
spacy
pdfplumber
python-docx
